\name{helix-package}
\alias{helix-package}
\alias{helix}
\docType{package}
\title{
\packageTitle{helix}
}
\description{
\packageDescription{helix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{helix}
Eine genaue Beschreibung folgt noch.
}
\author{
Harald Fiedler
Cavorit Consulting GmbH
}
\keyword{ package }
