% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/introMessage.R
\name{introMessage}
\alias{introMessage}
\title{introMessage}
\usage{
introMessage()
}
\description{
Druckt eine Nachricht aus.
}
\details{


Druckt eine Begrüßungs-Message aus.
}
\author{
Harald Fiedler
}

