% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/staninePlot.R
\name{staninePlot}
\alias{staninePlot}
\title{staninePlot}
\usage{
staninePlot()
}
\description{
Zeichnet stanine-Plot
}
\details{
Zeichnet eine Glockenkurve mit Stanine-Klassen in Cavorit-Farben
}
\author{
HF nach Felix Schönbrodt
}

