% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/staninePlotDichotom.R
\name{staninePlotDichotom}
\alias{staninePlotDichotom}
\title{staninePlotDichotom}
\usage{
staninePlotDichotom(Prozentrang, AlphaKanal = 1)
}
\arguments{
\item{Prozentrang}{Der Prozentrang, der abgezeichnet werden soll.}

\item{AlphaKanal}{Eine Zahl zwischen 0 und 1 für den Alphakanal}
}
\description{
Zeichnet Gaussglocke mit linker und rechter Farbe
}
\details{
Zeichnet eine Glockenkurve. Links des Prozentrangs wird alles in einer Farbe eingefäbrt. Rechts davon in einer anderen Farbe.
}
\author{
HF nach Felix Schönbrodt
}

